Removes shared code in PushAdapter, replaces AdapterLoader.load by loadAdapter
